feat : 데이터 형식 수정

+++ chat-da/src/main/java/site/chatda/domain/job/entity/JobSkillId.java
@@ -20,6 +20,6 @@ public class JobSkillId implements Serializable {
     private Integer jobId;
 
     @NotNull
-    @Column(columnDefinition = "TINYINT")
+    @Column(columnDefinition = "MEDIUMINT")
     private Integer skillId;
 }
